o/c/configcore,s/apparmor: mock aa root path  in prompting tests
